vizualizare in html a interogarii din BD; add si modelul de delete in html si view

+++ main/views.py
@@ -25,12 +25,9 @@ def persons(request):
     return render(request, "main/persons.html",{'persons':persons})
 
 def deleteperson(request, id):
-    #if request.method == 'GET':
-    print(id)
-    print(Person.objects.all())
-    Person.objects.all().filter(pk=int(id)).delete()
-    print(Person.objects.all())
-    persons = Person.objects.all()
+    if request.method == 'GET':
+        Person.objects.all().filter(pk=int(id)).delete()
+        persons = Person.objects.all()
     return render(request, 'main/persons.html', {'persons': persons})
 
 def person(request):
